fix: fix mistakes reviewed by my mentor
