+++ .gitignore
@@ -80,6 +80,7 @@ typings/
 
 # Nuxt.js build / generate output
 .nuxt
+.dist
 
 # Gatsby files
 .cache/
